Split up docker building steps and scripts
The Dockerfile is split into multiple RUN steps so it is easier to see
which one has failed.

The get_kustomize function was extracted into setup.sh because we now
run utils.sh with `set -e` so that errors are reported. setup.sh tries
to get some git branch information at the start, which is only available
when this image is actually used to build a component, and will cause
it's own Docker build to fail.